fix(pdp): select provider on ID, allow graceful fallback when no match (#195)
The registry allows the provider to have a different payee to provider address
so we can't match on it. But we can match on providerId because it can't
change.
Also allow for graceful fallback if we find data sets but none of them match
a provider we can use - so we'll just make a new one.